+++ src/createBEMElement.tsx
@@ -3,8 +3,9 @@ import * as PT from 'prop-types'
 import * as React from 'react'
 
 import { getDisplayName } from './helpers'
-import resolveBEMNode, { ReactRenderResult } from './resolveBEMNode'
-import { BEMElementClass, BEMElementProps } from './types'
+import { BEMBlockProviderContext } from './createBEMBlock'
+import resolveRenderedElement from './resolveRenderedElement'
+import { BEMElementProps } from './types'
 
 /**
  * Wraps a class with BEM element functionality, receiving the BEM block name
@@ -13,30 +14,32 @@ import { BEMElementClass, BEMElementProps } from './types'
  * @param ComponentClass The class to wrap with BEM element functionality.
  */
 export default function createBEMElement(
-	ComponentClass: BEMElementClass | React.ComponentClass,
+	ComponentClass: React.ComponentClass<any & BEMElementProps>,
 ) {
+	return class Wrapped extends ComponentClass {
 
-	return class Wrapped<P = {}, S = {}>
-	extends (ComponentClass as BEMElementClass)<P & BEMElementProps, S> {
+		static displayName =
+			`BEMElement(${getDisplayName(ComponentClass as any)})`
 
-		static displayName = `BEMElement(${getDisplayName(ComponentClass)})`
-
-		static contextTypes = {
-			...ComponentClass.contextTypes,
+		static contextTypes: BEMBlockProviderContext = {
+			...(ComponentClass.contextTypes || {}),
 			block: PT.string.isRequired,
 		}
 
+		context: BEMBlockProviderContext
+
+		static propTypes = {
+			...(ComponentClass.propTypes as any || {}),
+			element: PT.string,
+			modifiers: PT.any,
+		}
+
 		render() {
-			const {
-				element,
-				modifiers,
-			} = this.props
-			const rendered = super.render.call(this)
-			return rendered && resolveBEMNode(rendered, {
-				block: this.context.block,
-				element: element as string,
-				modifiers: [rendered.props.modifiers].concat([modifiers]),
-			}) as ReactRenderResult
+			return resolveRenderedElement(
+				super.render.call(this),
+				this.props,
+				this.context,
+			)
 		}
 	}
 }